Added some incomplete structures:
BrickGrid should take over the BrickInstance[][][] functionality in World.
BrickTreeNode is a tree structure where each node may depend on and be depended on by multiple other nodes. It's used to help determine which bricks can be dropped next
BrickStructure takes in an unsorted collection of bricks (which should represent one finished, non-floating brick structure) and builds a dependency-sorted collection.